enforce strong fetype

+++ core/src/db/SnapshotUtils.cpp
@@ -92,7 +92,7 @@ GetSnapshotIndex(const std::string& collection_name, const std::string& field_na
     auto field_elements = ss->GetFieldElementsByField(field_name);
     if (IsVectorField(field)) {
         for (auto& field_element : field_elements) {
-            if (field_element->GetFEtype() == (int64_t)milvus::engine::FieldElementType::FET_INDEX) {
+            if (field_element->GetFEtype() == milvus::engine::FieldElementType::FET_INDEX) {
                 index_info.index_name_ = field_element->GetName();
                 index_info.index_type_ = field_element->GetTypeName();
                 auto json = field_element->GetParams();
@@ -107,7 +107,7 @@ GetSnapshotIndex(const std::string& collection_name, const std::string& field_na
         }
     } else {
         for (auto& field_element : field_elements) {
-            if (field_element->GetFEtype() == (int64_t)milvus::engine::FieldElementType::FET_INDEX) {
+            if (field_element->GetFEtype() == milvus::engine::FieldElementType::FET_INDEX) {
                 index_info.index_name_ = field_element->GetName();
                 index_info.index_type_ = field_element->GetTypeName();
             }
