+++ core/src/db/meta/MetaResourceAttrs.h
@@ -114,7 +114,7 @@ AttrValue2Str(typename ResourceContext<ResourceT>::ResPtr src, const std::string
         int2str(int_value, value);
     } else if (F_FETYPE == attr) {
         auto fetype_field = std::dynamic_pointer_cast<snapshot::FEtypeField>(src);
-        int_value = fetype_field->GetFEtype();
+        int_value = (int)fetype_field->GetFEtype();
         int2str(int_value, value);
     } else if (F_FIELD_ID == attr) {
         auto field_id_field = std::dynamic_pointer_cast<snapshot::FieldIdField>(src);
